// packages/core/src/models/AddressModel.tsx
export class AddressModel {
    id?: number = undefined;
    tendiadiem?: string = undefined;
    loai?: string = undefined;
    tinhthanhpho?: string = undefined;
    quanhuyen?: string = undefined;
    phuongxa?: string = undefined;
    diachi?: string = undefined;
    nguoiphutrach?: string = undefined;
    sodienthoai?: string = undefined;
    trangthai?: string = undefined;
    ghichu?: string = undefined;
}